Move synthetic-data publisher test into publisher.rs unit tests
The `run_one_iteration_with_report` test fed a synthetic `Report::fake()`
through the publisher and asserted on the exported `ResourceMetrics`. It had
no dependency on the global `nm` registry, so there was no reason to host
it in a separate integration-test binary — that only slowed builds and
added a redundant link of the lib + opentelemetry stack.

Move it into `publisher.rs::tests` alongside the other publisher unit tests
and delete the integration-test file. The remaining integration tests under
`tests/` either touch the global nm registry (`run_one_iteration`,
`run_one_iteration_deltas`, `publish_forever`) or install a custom
`#[global_allocator]` (`histogram_deltas_alloc`), so they must stay as
separate binaries.
